Added dockerfile for puppeteer and created a new file for puppeteer testing.
