+++ Dockerfile
@@ -1,0 +1,12 @@
+FROM ghcr.io/puppeteer/puppeteer:19.10.0
+
+ENV PUPPETEER_SKIP_CHROMIUM_DOWNLOAD=true \
+    PUPPETEER_EXECUTABLE_PATH=/usr/bin/google-chrome-stable
+
+WORKDIR /usr/src/app
+
+COPY package*.json ./
+RUN npm ci
+COPY . .
+
+CMD [ "node", "index" ]
