+++ package.json
@@ -24,6 +24,7 @@
     "express": "^4.18.2",
     "jsonwebtoken": "^9.0.0",
     "mysql2": "^3.2.3",
+    "puppeteer": "^19.10.0",
     "puppeteer-core": "^6.0.0",
     "puppeteer-page-proxy": "^1.3.0",
     "puppeteer-proxy": "^2.1.2"
